+++ crates/pallets/starknet/src/lib.rs
@@ -82,7 +82,7 @@ use mp_digest_log::MADARA_ENGINE_ID;
 use mp_starknet::block::{Block as StarknetBlock, BlockTransactions, Header as StarknetHeader, MaxTransactions};
 use mp_starknet::crypto::commitment;
 use mp_starknet::crypto::hash::pedersen::PedersenHasher;
-use mp_starknet::execution::{
+use mp_starknet::execution::types::{
     CallEntryPointWrapper, ClassHashWrapper, ContractAddressWrapper, ContractClassWrapper, EntryPointTypeWrapper,
 };
 use mp_starknet::storage::{StarknetStorageSchemaVersion, PALLET_STARKNET_SCHEMA};
@@ -439,6 +439,10 @@ pub mod pallet {
                 }
                 Err(e) => {
                     log!(error, "Transaction execution failed: {:?}", e);
+                    #[cfg(feature = "std")]
+                    use std::println;
+                    #[cfg(feature = "std")]
+                    println!("Invoke error {:?}", e);
                     return Err(Error::<T>::TransactionExecutionFailed.into());
                 }
             };
