+++ crates/primitives/starknet/src/execution/mod.rs
@@ -1,381 +1,17 @@
 //! Starknet execution functionality.
 
-/// Types related to entrypoints.
-pub mod types;
+mod call_entrypoint_wrapper;
+mod contract_class_wrapper;
+mod entrypoint_wrapper;
 
-use alloc::collections::BTreeMap;
-use alloc::sync::Arc;
-use alloc::vec::Vec;
-use alloc::{format, vec};
+/// All the types related to the execution of a transaction.
+pub mod types {
+    /// Type wrapper for a contract address.
+    pub type ContractAddressWrapper = [u8; 32];
 
-use blockifier::block_context::BlockContext;
-use blockifier::execution::contract_class::ContractClass;
-use blockifier::execution::entry_point::{CallEntryPoint, CallInfo, CallType, ExecutionContext, ExecutionResources};
-use blockifier::execution::execution_utils::{cairo_vm_program_to_sn_api, sn_api_to_cairo_vm_program};
-use blockifier::state::state_api::State;
-use blockifier::transaction::objects::AccountTransactionContext;
-use cairo_vm::types::errors::program_errors::ProgramError;
-use frame_support::{BoundedBTreeMap, BoundedVec};
-#[cfg(feature = "std")]
-use frame_support::{Deserialize, Serialize};
-use serde::de::Error as DeserializationError;
-use serde::{Deserializer, Serializer};
-use serde_json::{from_slice, to_string};
-use sp_core::{ConstU32, H256, U256};
-use starknet_api::api_core::{ClassHash, ContractAddress, EntryPointSelector};
-use starknet_api::deprecated_contract_class::{
-    EntryPoint, EntryPointOffset, EntryPointType, Program as DeprecatedProgram,
-};
-use starknet_api::hash::StarkFelt;
-use starknet_api::stdlib::collections::HashMap;
-use starknet_api::transaction::Calldata;
-use thiserror_no_std::Error;
-
-use self::types::{EntryPointExecutionErrorWrapper, EntryPointExecutionResultWrapper};
-use crate::block::serialize::SerializeBlockContext;
-use crate::block::Block as StarknetBlock;
-use crate::transaction::types::MaxArraySize;
-
-/// The address of a contract.
-pub type ContractAddressWrapper = [u8; 32];
-
-/// Maximum vector sizes.
-// type MaxCalldataSize = ConstU32<4294967295>;
-// type MaxAbiSize = ConstU32<4294967295>;
-type MaxProgramSize = ConstU32<4294967295>;
-/// Max number of entrypoints.
-pub type MaxEntryPoints = ConstU32<4294967295>;
-type MaxEntryPointsType = ConstU32<3>;
-
-/// Wrapper type for class hash field.
-pub type ClassHashWrapper = [u8; 32];
-
-/// Contract Class
-#[derive(
-    Clone,
-    Debug,
-    PartialEq,
-    Eq,
-    scale_codec::Encode,
-    scale_codec::Decode,
-    scale_info::TypeInfo,
-    Default,
-    scale_codec::MaxEncodedLen,
-)]
-#[cfg_attr(feature = "std", derive(serde::Serialize, serde::Deserialize))]
-pub struct ContractClassWrapper {
-    /// Contract class program json.
-    pub program: BoundedVec<u8, MaxProgramSize>,
-    /// Contract class entrypoints.
-    #[cfg_attr(
-        feature = "std",
-        serde(deserialize_with = "deserialize_entrypoints", serialize_with = "serialize_entrypoints")
-    )]
-    pub entry_points_by_type:
-        BoundedBTreeMap<EntryPointTypeWrapper, BoundedVec<EntryPointWrapper, MaxEntryPoints>, MaxEntryPointsType>,
-}
-/// MOI ETRE GENIE
-#[cfg(feature = "std")]
-pub fn deserialize_entrypoints<'de, D: Deserializer<'de>>(
-    deserializer: D,
-) -> Result<
-    BoundedBTreeMap<EntryPointTypeWrapper, BoundedVec<EntryPointWrapper, MaxEntryPoints>, MaxEntryPointsType>,
-    D::Error,
-> {
-    let entrypoints = BTreeMap::deserialize(deserializer)?;
-    BoundedBTreeMap::try_from(entrypoints)
-        .map_err(|_| DeserializationError::custom("Couldn't convert BTreeMap to BoundedBTreeMap".to_string()))
-}
-
-/// Converts the program type from SN API into a Cairo VM-compatible type.
-#[cfg(feature = "std")]
-pub fn serialize_entrypoints<S: Serializer>(
-    v: &BoundedBTreeMap<EntryPointTypeWrapper, BoundedVec<EntryPointWrapper, MaxEntryPoints>, MaxEntryPointsType>,
-    serializer: S,
-) -> Result<S::Ok, S::Error> {
-    v.clone().into_inner().serialize(serializer)
-}
-impl ContractClassWrapper {
-    /// Creates a new instance of a contract class.
-    pub fn new(
-        program: BoundedVec<u8, MaxProgramSize>,
-        entry_points_by_type: BoundedBTreeMap<
-            EntryPointTypeWrapper,
-            BoundedVec<EntryPointWrapper, MaxEntryPoints>,
-            MaxEntryPointsType,
-        >,
-    ) -> Self {
-        Self { program, entry_points_by_type }
-    }
-}
-
-/// Some doc
-#[derive(Debug, Error)]
-pub enum ContractClassFromWrapperError {
-    /// Other doc
-    #[error(transparent)]
-    Program(#[from] ProgramError),
-    /// ANOTHER DOC
-    #[error(transparent)]
-    Serde(#[from] serde_json::Error),
-}
-
-impl TryFrom<ContractClassWrapper> for ContractClass {
-    type Error = ContractClassFromWrapperError;
-
-    fn try_from(wrapper: ContractClassWrapper) -> Result<Self, Self::Error> {
-        let mut entrypoints = HashMap::new();
-        wrapper.entry_points_by_type.into_iter().for_each(|(key, val)| {
-            entrypoints.insert(key.into(), val.iter().map(|entrypoint| EntryPoint::from(entrypoint.clone())).collect());
-        });
-        let program = from_slice::<DeprecatedProgram>(wrapper.program.as_ref())?;
-        let program = sn_api_to_cairo_vm_program(program)?;
-        Ok(ContractClass { program, entry_points_by_type: entrypoints })
-    }
-}
-
-impl TryFrom<ContractClass> for ContractClassWrapper {
-    type Error = serde_json::Error;
-
-    fn try_from(contract_class: ContractClass) -> Result<Self, Self::Error> {
-        let program = cairo_vm_program_to_sn_api(contract_class.program)?;
-        let program_string = to_string(&program).unwrap();
-        let mut entrypoints = BTreeMap::new();
-        for (key, val) in contract_class.entry_points_by_type.iter() {
-            entrypoints.insert(
-                (*key).into(),
-                BoundedVec::try_from(val.iter().map(|elt| elt.clone().into()).collect::<Vec<EntryPointWrapper>>())
-                    .unwrap(),
-            );
-        }
-        Ok(Self {
-            program: BoundedVec::try_from(program_string.as_bytes().to_vec()).unwrap(),
-            entry_points_by_type: BoundedBTreeMap::try_from(entrypoints).unwrap(),
-        })
-    }
-}
-
-/// Enum that represents all the entrypoints types.
-#[derive(
-    Clone,
-    Debug,
-    PartialEq,
-    Eq,
-    scale_codec::Encode,
-    scale_codec::Decode,
-    scale_info::TypeInfo,
-    scale_codec::MaxEncodedLen,
-    PartialOrd,
-    Ord,
-    Hash,
-)]
-#[cfg_attr(feature = "std", derive(serde::Serialize, serde::Deserialize))]
-pub enum EntryPointTypeWrapper {
-    /// Constructor.
-    Constructor,
-    /// External.
-    External,
-    /// L1 Handler.
-    L1Handler,
-}
-
-impl From<EntryPointType> for EntryPointTypeWrapper {
-    fn from(entry_point_type: EntryPointType) -> Self {
-        match entry_point_type {
-            EntryPointType::Constructor => EntryPointTypeWrapper::Constructor,
-            EntryPointType::External => EntryPointTypeWrapper::External,
-            EntryPointType::L1Handler => EntryPointTypeWrapper::L1Handler,
-        }
-    }
-}
-
-impl From<EntryPointTypeWrapper> for EntryPointType {
-    fn from(entrypoint: EntryPointTypeWrapper) -> Self {
-        match entrypoint {
-            EntryPointTypeWrapper::Constructor => EntryPointType::Constructor,
-            EntryPointTypeWrapper::External => EntryPointType::External,
-            EntryPointTypeWrapper::L1Handler => EntryPointType::L1Handler,
-        }
-    }
-}
-
-impl EntryPointTypeWrapper {
-    /// Convert to starknet entrypoint type.
-    pub fn to_starknet_entry_point_type(&self) -> EntryPointType {
-        match self {
-            EntryPointTypeWrapper::Constructor => EntryPointType::Constructor,
-            EntryPointTypeWrapper::External => EntryPointType::External,
-            EntryPointTypeWrapper::L1Handler => EntryPointType::L1Handler,
-        }
-    }
-}
-
-/// Representation of a Starknet Entry Point.
-#[derive(
-    Clone,
-    Debug,
-    PartialEq,
-    Eq,
-    scale_codec::Encode,
-    scale_codec::Decode,
-    scale_info::TypeInfo,
-    scale_codec::MaxEncodedLen,
-    PartialOrd,
-    Ord,
-)]
-#[cfg_attr(feature = "std", derive(serde::Serialize, serde::Deserialize))]
-pub struct EntryPointWrapper {
-    /// The entrypoint selector
-    pub entrypoint_selector: H256,
-    /// The entrypoint offset
-    pub entrypoint_offset: u128,
-}
-
-impl EntryPointWrapper {
-    /// Creates a new instance of an entrypoint.
-    pub fn new(entrypoint_selector: H256, entrypoint_offset: u128) -> Self {
-        Self { entrypoint_selector, entrypoint_offset }
-    }
-}
-
-impl From<EntryPoint> for EntryPointWrapper {
-    fn from(entry_point: EntryPoint) -> Self {
-        Self {
-            entrypoint_selector: H256::from_slice(entry_point.selector.0.bytes()),
-            entrypoint_offset: entry_point.offset.0 as u128,
-        }
-    }
-}
-
-impl From<EntryPointWrapper> for EntryPoint {
-    fn from(entry_point: EntryPointWrapper) -> Self {
-        Self {
-            selector: EntryPointSelector(StarkFelt(entry_point.entrypoint_selector.to_fixed_bytes())),
-            offset: EntryPointOffset(entry_point.entrypoint_offset as usize),
-        }
-    }
-}
-
-/// Representation of a Starknet Call Entry Point.
-#[derive(
-    Clone,
-    Debug,
-    PartialEq,
-    Eq,
-    scale_codec::Encode,
-    scale_codec::Decode,
-    scale_info::TypeInfo,
-    scale_codec::MaxEncodedLen,
-)]
-#[cfg_attr(feature = "std", derive(serde::Serialize, serde::Deserialize))]
-pub struct CallEntryPointWrapper {
-    /// The class hash
-    pub class_hash: Option<ClassHashWrapper>,
-    /// The entrypoint type
-    pub entrypoint_type: EntryPointTypeWrapper,
-    /// The entrypoint selector
-    /// An invoke transaction without an entry point selector invokes the 'execute' function.
-    pub entrypoint_selector: Option<H256>,
-    /// The Calldata
-    pub calldata: BoundedVec<U256, MaxArraySize>,
-    /// The storage address
-    pub storage_address: ContractAddressWrapper,
-    /// The caller address
-    pub caller_address: ContractAddressWrapper,
-}
-impl EntryPointTypeWrapper {
-    /// Convert Madara entrypoint type to Starknet entrypoint type.
-    pub fn to_starknet(&self) -> EntryPointType {
-        match self {
-            Self::Constructor => EntryPointType::Constructor,
-            Self::External => EntryPointType::External,
-            Self::L1Handler => EntryPointType::L1Handler,
-        }
-    }
-}
-
-impl CallEntryPointWrapper {
-    /// Creates a new instance of a call entrypoint.
-    pub fn new(
-        class_hash: Option<ClassHashWrapper>,
-        entrypoint_type: EntryPointTypeWrapper,
-        entrypoint_selector: Option<H256>,
-        calldata: BoundedVec<U256, MaxArraySize>,
-        storage_address: ContractAddressWrapper,
-        caller_address: ContractAddressWrapper,
-    ) -> Self {
-        Self { class_hash, entrypoint_type, entrypoint_selector, calldata, storage_address, caller_address }
-    }
-
-    /// Convert to Starknet CallEntryPoint
-    pub fn to_starknet_call_entry_point(&self) -> CallEntryPoint {
-        let class_hash = self.class_hash.map(|class_hash| ClassHash(StarkFelt::new(class_hash).unwrap()));
-        CallEntryPoint {
-            class_hash,
-            entry_point_type: self.entrypoint_type.to_starknet(),
-            entry_point_selector: EntryPointSelector(
-                StarkFelt::new(self.entrypoint_selector.unwrap_or_default().0).unwrap(),
-            ),
-            calldata: Calldata(Arc::new(
-                self.calldata
-                    .clone()
-                    .into_inner()
-                    .iter()
-                    .map(|x| StarkFelt::try_from(format!("0x{x:X}").as_str()).unwrap())
-                    .collect(),
-            )),
-            storage_address: ContractAddress::try_from(StarkFelt::new(self.storage_address).unwrap()).unwrap(),
-            caller_address: ContractAddress::try_from(StarkFelt::new(self.caller_address).unwrap()).unwrap(),
-            call_type: CallType::Call,
-            // I have no idea what I'm doing
-            // starknet-lib is constantly breaking it's api
-            // I hope it's nothing important ¯\_(ツ)_/¯
-            code_address: None,
-        }
-    }
-
-    /// Executes an entry point.
-    ///
-    /// # Arguments
-    ///
-    /// * `self` - The entry point to execute.
-    /// * `state` - The state to execute the entry point on.
-    /// * `block` - The block to execute the entry point on.
-    /// * `fee_token_address` - The fee token address.
-    ///
-    /// # Returns
-    ///
-    /// * The result of the entry point execution.
-    pub fn execute<S: State>(
-        &self,
-        state: &mut S,
-        block: StarknetBlock,
-        fee_token_address: ContractAddressWrapper,
-    ) -> EntryPointExecutionResultWrapper<CallInfo> {
-        let call_entry_point = self.to_starknet_call_entry_point();
-
-        let execution_resources = &mut ExecutionResources::default();
-        let execution_context = &mut ExecutionContext::default();
-        let account_context = AccountTransactionContext::default();
-
-        // Create the block context.
-        let block_context = BlockContext::try_serialize(block.header().clone(), fee_token_address)
-            .map_err(|_| EntryPointExecutionErrorWrapper::BlockContextSerializationError)?;
-
-        call_entry_point
-            .execute(state, execution_resources, execution_context, &block_context, &account_context)
-            .map_err(EntryPointExecutionErrorWrapper::EntryPointExecution)
-    }
-}
-impl Default for CallEntryPointWrapper {
-    fn default() -> Self {
-        Self {
-            class_hash: Some(ClassHashWrapper::default()),
-            entrypoint_type: EntryPointTypeWrapper::External,
-            entrypoint_selector: Some(H256::default()),
-            calldata: BoundedVec::try_from(vec![U256::zero(); 32]).unwrap(),
-            storage_address: ContractAddressWrapper::default(),
-            caller_address: ContractAddressWrapper::default(),
-        }
-    }
+    /// Wrapper type for class hash field.
+    pub type ClassHashWrapper = [u8; 32];
+    pub use super::call_entrypoint_wrapper::*;
+    pub use super::contract_class_wrapper::*;
+    pub use super::entrypoint_wrapper::*;
 }
